refactor(ts-sdk): split client.ts + types.ts into sub-modules ≤500L (B.35)
Sprint 2 Wave 8 item B.35: client.ts (1038→422L) split into client-core,
search-methods, graph-methods, validation. types.ts (1088→barrel) split
into core, search, graph, agent, query, index-types, backend, errors,
endpoints. All imports preserved via barrel re-exports.
