+++ sdks/typescript/src/client/index.ts
@@ -1,0 +1,8 @@
+/**
+ * VelesDB Client - barrel re-export
+ * @packageDocumentation
+ */
+
+export { requireNonEmptyString, requireVector, validateDocsBatch, validateDocument, validateRestPointId } from './validation';
+export * from './graph-methods';
+export * from './search-methods';
