+++ sdks/typescript/src/types/index.ts
@@ -1,0 +1,17 @@
+/**
+ * VelesDB TypeScript SDK - Type Definitions (barrel re-export)
+ *
+ * All type sub-modules are re-exported here so that consumers can
+ * import from `./types` or `../types` exactly as before.
+ * @packageDocumentation
+ */
+
+export * from './core';
+export * from './search';
+export * from './graph';
+export * from './agent';
+export * from './query';
+export * from './index-types';
+export * from './backend';
+export * from './errors';
+export * from './endpoints';
